+++ src/encoder.rs
@@ -6,6 +6,7 @@
 //!
 use std::vec::Vec;
 
+use crate::Error;
 use crate::geom::{Transform, Vec2};
 
 #[derive(Clone, Debug)]
@@ -35,16 +36,18 @@ pub enum GeomType {
     Polygon,
 }
 
-/// Encoder for Point, Linestring or Polygon geometry.
+/// Encoder for [Feature](struct.Feature.html) geometry.  This can consist of
+/// Point, Linestring or Polygon data.
 ///
 /// # Example
 /// ```
-/// use mvt::{GeomEncoder,GeomType,Transform};
-/// let mut encoder = GeomEncoder::new(GeomType::Point, Transform::new());
-/// encoder.add_point(0.0, 0.0);
-/// encoder.add_point(10.0, 0.0);
-/// // ...
-/// // create a feature
+/// # use mvt::{Error,GeomEncoder,GeomType,Transform};
+/// # fn main() -> Result<(), Error> {
+/// let geom_data = GeomEncoder::new(GeomType::Point, Transform::new())
+///                             .add_point(0.0, 0.0)
+///                             .add_point(10.0, 0.0)
+///                             .encode()?;
+/// # Ok(()) }
 /// ```
 pub struct GeomEncoder {
     geom_tp: GeomType,
@@ -53,6 +56,24 @@ pub struct GeomEncoder {
     y: i32,
     cmd_offset: usize,
     count: u32,
+    data: Vec<u32>,
+}
+
+/// Validated geometry data for [Feature](struct.Feature.html)s.  Use
+/// [GeomEncoder](struct.GeomEncoder.html) to encode.
+///
+/// # Example
+/// ```
+/// # use mvt::{Error,GeomEncoder,GeomType,Transform};
+/// # fn main() -> Result<(), Error> {
+/// let geom_data = GeomEncoder::new(GeomType::Point, Transform::new())
+///                             .add_point(0.0, 0.0)
+///                             .add_point(10.0, 0.0)
+///                             .encode()?;
+/// # Ok(()) }
+/// ```
+pub struct GeomData {
+    geom_tp: GeomType,
     data: Vec<u32>,
 }
 
@@ -91,11 +112,6 @@ impl GeomEncoder {
         }
     }
 
-    /// Get the geometry type
-    pub(crate) fn geom_type(&self) -> GeomType {
-        self.geom_tp.clone()
-    }
-
     /// Add a Command
     fn command(&mut self, cmd: Command, count: u32) {
         self.cmd_offset = self.data.len();
@@ -109,8 +125,8 @@ impl GeomEncoder {
         self.data[off] = CommandInt::new(cmd, count).encode();
     }
 
-    /// Encode one point with relative coörindates.
-    fn encode_point(&mut self, x: f64, y: f64) {
+    /// Push one point with relative coörindates.
+    fn push_point(&mut self, x: f64, y: f64) {
         let p = self.transform * Vec2::new(x, y);
         let x = p.x as i32;
         let y = p.y as i32;
@@ -124,7 +140,7 @@ impl GeomEncoder {
     }
 
     /// Add a point.
-    pub fn add_point(&mut self, x: f64, y: f64) {
+    pub fn add_point(mut self, x: f64, y: f64) -> Self {
         match self.geom_tp {
             GeomType::Point => {
                 if self.count == 0 {
@@ -142,12 +158,15 @@ impl GeomEncoder {
                 _ => (),
             },
         }
-        self.encode_point(x, y);
+        self.push_point(x, y);
         self.count += 1;
+        self
     }
 
     /// Complete the current geometry (for multilinestring / multipolygon).
-    pub fn complete_geom(&mut self) {
+    pub fn complete_geom(mut self) -> Result<Self, Error> {
+        // FIXME: return Error::InvalidGeometry
+        //        if "MUST" rules in the spec are violated
         match self.geom_tp {
             GeomType::Point => (),
             GeomType::Linestring => {
@@ -164,17 +183,41 @@ impl GeomEncoder {
                 self.count = 0;
             }
         }
+        Ok(self)
     }
 
-    /// Get the encoded data.
-    pub(crate) fn to_vec(mut self) -> Vec<u32> {
-        if let GeomType::Point = self.geom_tp {
+    /// Encode the geometry data, consuming the encoder.
+    pub fn encode(mut self) -> Result<GeomData, Error> {
+        // FIXME: return Error::InvalidGeometry
+        //        if "MUST" rules in the spec are violated
+        self = if let GeomType::Point = self.geom_tp {
             if self.count > 1 {
                 self.set_command(Command::MoveTo, self.count);
             }
+            self
         } else {
-            self.complete_geom();
-        }
+            self.complete_geom()?
+        };
+        Ok(GeomData::new(self.geom_tp, self.data))
+    }
+}
+
+impl GeomData {
+    /// Create new geometry data.
+    ///
+    /// * `geom_tp` Geometry type.
+    /// * `data` Validated geometry.
+    fn new(geom_tp: GeomType, data: Vec<u32>) -> Self {
+        GeomData { geom_tp, data }
+    }
+
+    /// Get the geometry type
+    pub(crate) fn geom_type(&self) -> GeomType {
+        self.geom_tp.clone()
+    }
+
+    /// Get the geometry data
+    pub(crate) fn into_vec(self) -> Vec<u32> {
         self.data
     }
 }
@@ -185,69 +228,87 @@ mod test {
     // Examples from MVT spec:
     #[test]
     fn test_point() {
-        let mut pe = GeomEncoder::new(GeomType::Point, Transform::new());
-        pe.add_point(25.0, 17.0);
-        let b = pe.to_vec();
-        assert_eq!(b, vec!(9, 50, 34));
+        let v = GeomEncoder::new(GeomType::Point, Transform::new())
+                            .add_point(25.0, 17.0)
+                            .encode()
+                            .unwrap()
+                            .into_vec();
+        assert_eq!(v, vec!(9, 50, 34));
     }
     #[test]
     fn test_multipoint() {
-        let mut pe = GeomEncoder::new(GeomType::Point, Transform::new());
-        pe.add_point(5.0, 7.0);
-        pe.add_point(3.0, 2.0);
-        let b = pe.to_vec();
-        assert_eq!(b, vec!(17, 10, 14, 3, 9));
+        let v = GeomEncoder::new(GeomType::Point, Transform::new())
+                            .add_point(5.0, 7.0)
+                            .add_point(3.0, 2.0)
+                            .encode()
+                            .unwrap()
+                            .into_vec();
+        assert_eq!(v, vec!(17, 10, 14, 3, 9));
     }
     #[test]
     fn test_linestring() {
-        let mut le = GeomEncoder::new(GeomType::Linestring, Transform::new());
-        le.add_point(2.0, 2.0);
-        le.add_point(2.0, 10.0);
-        le.add_point(10.0, 10.0);
-        let b = le.to_vec();
-        assert_eq!(b, vec!(9, 4, 4, 18, 0, 16, 16, 0));
+        let v = GeomEncoder::new(GeomType::Linestring, Transform::new())
+                            .add_point(2.0, 2.0)
+                            .add_point(2.0, 10.0)
+                            .add_point(10.0, 10.0)
+                            .encode()
+                            .unwrap()
+                            .into_vec();
+        assert_eq!(v, vec!(9, 4, 4, 18, 0, 16, 16, 0));
     }
     #[test]
     fn test_multilinestring() {
-        let mut le = GeomEncoder::new(GeomType::Linestring, Transform::new());
-        le.add_point(2.0, 2.0);
-        le.add_point(2.0, 10.0);
-        le.add_point(10.0, 10.0);
-        le.complete_geom();
-        le.add_point(1.0, 1.0);
-        le.add_point(3.0, 5.0);
-        let b = le.to_vec();
-        assert_eq!(b, vec!(9, 4, 4, 18, 0, 16, 16, 0, 9, 17, 17, 10, 4, 8));
+        let v = GeomEncoder::new(GeomType::Linestring, Transform::new())
+                            .add_point(2.0, 2.0)
+                            .add_point(2.0, 10.0)
+                            .add_point(10.0, 10.0)
+                            .complete_geom()
+                            .unwrap()
+                            .add_point(1.0, 1.0)
+                            .add_point(3.0, 5.0)
+                            .encode()
+                            .unwrap()
+                            .into_vec();
+        assert_eq!(v, vec!(9, 4, 4, 18, 0, 16, 16, 0, 9, 17, 17, 10, 4, 8));
     }
     #[test]
     fn test_polygon() {
-        let mut pe = GeomEncoder::new(GeomType::Polygon, Transform::new());
-        pe.add_point(3.0, 6.0);
-        pe.add_point(8.0, 12.0);
-        pe.add_point(20.0, 34.0);
-        let b = pe.to_vec();
-        assert_eq!(b, vec!(9, 6, 12, 18, 10, 12, 24, 44, 15));
+        let v = GeomEncoder::new(GeomType::Polygon, Transform::new())
+                            .add_point(3.0, 6.0)
+                            .add_point(8.0, 12.0)
+                            .add_point(20.0, 34.0)
+                            .encode()
+                            .unwrap()
+                            .into_vec();
+        assert_eq!(v, vec!(9, 6, 12, 18, 10, 12, 24, 44, 15));
     }
     #[test]
     fn test_multipolygon() {
-        let mut pe = GeomEncoder::new(GeomType::Polygon, Transform::new());
-        pe.add_point(0.0, 0.0);
-        pe.add_point(10.0, 0.0);
-        pe.add_point(10.0, 10.0);
-        pe.add_point(0.0, 10.0);
-        pe.complete_geom(); // positive area => exterior ring
-        pe.add_point(11.0, 11.0);
-        pe.add_point(20.0, 11.0);
-        pe.add_point(20.0, 20.0);
-        pe.add_point(11.0, 20.0);
-        pe.complete_geom(); // positive area => exterior ring
-        pe.add_point(13.0, 13.0);
-        pe.add_point(13.0, 17.0);
-        pe.add_point(17.0, 17.0);
-        pe.add_point(17.0, 13.0);
-        let b = pe.to_vec(); // negative area => interior ring
+        let v = GeomEncoder::new(GeomType::Polygon, Transform::new())
+                            // positive area => exterior ring
+                            .add_point(0.0, 0.0)
+                            .add_point(10.0, 0.0)
+                            .add_point(10.0, 10.0)
+                            .add_point(0.0, 10.0)
+                            .complete_geom()
+                            .unwrap()
+                            // positive area => exterior ring
+                            .add_point(11.0, 11.0)
+                            .add_point(20.0, 11.0)
+                            .add_point(20.0, 20.0)
+                            .add_point(11.0, 20.0)
+                            .complete_geom()
+                            .unwrap()
+                            // negative area => interior ring
+                            .add_point(13.0, 13.0)
+                            .add_point(13.0, 17.0)
+                            .add_point(17.0, 17.0)
+                            .add_point(17.0, 13.0)
+                            .encode()
+                            .unwrap()
+                            .into_vec();
         assert_eq!(
-            b,
+            v,
             vec!(
                 9, 0, 0, 26, 20, 0, 0, 20, 19, 0, 15, 9, 22, 2, 26, 18, 0, 0, 18, 17, 0, 15, 9, 4,
                 13, 26, 0, 8, 8, 0, 0, 7, 15