+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ## [Unreleased]
 ### Added
 * Check extent when adding layer to tile
+* GeomEncoder now has encode method to create GeomData struct
+* New error variant: InvalidGeometry
 
 ### Changed
+* GeomEncoder now uses builder pattern
 * Made Tile::compute_size private
 
 ## [0.1.0] - 2010-01-10